feat: update schema and translation from schema to Config
The older method of parsing the schema into a json patch template,
then patching it on top of a base config has now been discontinued in
favour of directly assigning to talos's v1alpha1.Config type and
generating the config yaml from that. This reduces the opportunity for
bugs that would arise from a badly formatted template or user input,
and opens up the provider to expose the full range of configuration
options Talos provides.

In addition common schema elements and helpers to translate between
them and go data types have been moved to their own common.go file.

This commit also adds validation in some resources for some of the new
fields to ensure user errors are caught before being sent to the Talos
Config.