buid: add acceptance test target to Makefile
The acceptance tests require the node http-server program to be
installed. This is to speed up virtual machine creation as each one
will attempt to grab the talos iso. When running the test a local
registry must also be set up according to the instructions on the
Talos website link:
"https://www.talos.dev/v1.0/talos-guides/configuration/pull-through-cache/"