feat: update provider to terraform plugin framework
This is a very large commit that completely rewrites the plugin to use
the actively developed terraform plugin framework. This is because the
framework is designed with highly nested schema in mind, which Talos's
configuration approach requires. The approach used by the framework
makes implementing generic methods for moving data in and out of
controlplane and worker nodes much harder, so an interface approach
was taken. This does weaken type guarentees at the boundry between
terraform and Talos's data and will be a likely source of bugs in the
future. A generics based approach would work better if feasable. Or a
type reflection approach that gets values from the resource's structs
using the "tfsdk" tags that need to be included.

Tests, acceptance and unit, have not been updated. Validation of
fields has also been lost for the time being. Reading resource state
has also not been implemented.