combined commit including import, schema changes and fixes
This commit is lazy, but a consequence of hacking away at a new
framework. Large scale refactoring and unsatisfactory git discipline
resulted in me bundling these changes into one commit and calling it a
day. As an overview this commit.
+ Change file imports to work account for refactored/added code
+ Fix udev's schema type being incorrect (map to list)
+ Fix kernel argument schema (map to list)
+ Add proxy schema to the worker node resource
+ Remove bootstrap and bootstrap IP from worker node and replace with
"config_ip". Will be replaced with a field in the "device" config so a
specific interface is used for configuration regardless of IP.
+ Return struct references from worker TalosData.
+ Force InstallBootloader to true for now.