feat: reimplemnent controlplane node deletion
Controlplane deletion has been re-added to the code. Still has the
same issues as before (no logic to check whether the node should be
gracefully or forecefully deleted). Ideally it would gracefully reset
the node if it's part of a cluster and forcefully reset when the
entire cluster is being destroyed.