refactor: move code to datatypes package
Code that was originally in the "common.go" file has now been moved to
a seperate package. Each data struct has an accompanying example that
can be used to generate documentation or in tests. Additionally
interface implementations are in individual files to make things
easier. Lastly there has been some changes to schema and test
behaviour that was unavoidable during these last changes.