feat: move MAC based UP lookup to seperate provider
The MAC lookup code was prone to bugs and made acceptance testing
significantly harder to reproduce reliably. It was moved to a
seperate provider which has much better testing. It simplifies
resource creation code, simplifies it's schema and usage. It's the
first step to overhauling the acceptance testing to run as a Github
Action.