test: overhaul acceptance testing
Acceptance testing now runs as user mode VMs inside a user network
namespace. This provides full isolation from system network
configuration and enables testing inside Github actions. The speed of
running inside a Github action is lacking due to the absence of KVM
acceleration. When ran on a local machine it is enabled.