refactor: move from data interface to using a list of functions
The previous approach of converting data to and from Talos datatypes
was cumbersome and required a lot of code in each resource. This new
approach only requires a list of provider datatypes, which implement
the ConfigDataFunc interface. This list is passed to a function that
applies these functions to the Talos state, using data from the plan's
state.

The next step would be to pass child datafuncs for any child
struct.