feat: full support for talos node creation options
The provider now passes the full test set that checks whether, from a
set of terraform data, a valid talos node configuration struct
containing all possible options can be created. A single issue with
setting the BondUseCarrier value was introduced and will be fixed
soon.

The next step is migrating all existing Data() functions to DataFunc()
and then refactoring DataFunc() to Data().

After that a similar set of ReadFunc()s will be created, with the same
refactoring done once it passes read testing.

There is still some work to be done making the DataFunc helper
functions safer, such as always returning an error if passed a null
pointer as the destination.