refactor: add helpers for reading and setting terraform types
Terraform values are nullable, while Talos expects null values to not
be set. This leads to difficulty when creating and reading
resources. In this case it's alleviated by creating a standard
optional type that lets the read and data funcs easily determine if a
value is unset, and to set the appropriate value in a way expected by
both Terraform and Talos. Although only implemented for three
primitive types, this approach can be expanded to more datatypes and
will likely be what the other helpers are refactored to.