feat: generate node fields from base configuration
The cluster configuration resource can define basic shared parameters
for node creation. Prior to this commit these were not factored into
the node's Terraform state, resulting in errors during usage. This is
an initial attempt at implementing this, it is incomplete and needs
testcases to be created. Conflicts between the cluster and node
configuration need to be reconciled, most likely by node values
overriding cluster values.