feat: implement ReadFuncs for network related settings
Readfuncs for network interface lists and the missing aspects of the
TalosNetworkConfig stub were implemented. The kubespan ReadFunc was
also implemented. This particular area of the code adopts a hybrid
approach of a ReadFunc for the top level resource and regular
functions for everything down the stack. This is the simplest option
and will be the target when the other ReadFuncs are refactored.