feat: implement VolumeMount ReadFunc for different parent types
The design of the ReadFunc approach does not allow us to inject any
information of the type's parent. This leads to issues when reading
nested data inside slices, like VolumeMounts. The naive solution is to
create a ReadFunc for each individual parent type. The simpler and
better option is to create a function and use that.

TODO: Refactor this